# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Clean Architecture Example with Go, Docker.",
        "title": "Payment System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/merchants/bybusiness/{businessID}": {
            "get": {
                "description": "Retrieve all merchants belonging to a specific Business",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchants"
                ],
                "summary": "List Merchants by Business",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Business UUID",
                        "name": "businessID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Merchant"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchants/delete/{id}": {
            "delete": {
                "description": "Logic delete of a merchant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchants"
                ],
                "summary": "Remove a Merchant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Merchant UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchants/new": {
            "post": {
                "description": "Creates a new merchant linked to a specific business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchants"
                ],
                "summary": "Register a new Merchant",
                "parameters": [
                    {
                        "description": "Merchant Registration",
                        "name": "merchant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler.createMerchantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Merchant"
                        }
                    },
                    "400": {
                        "description": "Invalid input or Business UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchants/{id}": {
            "get": {
                "description": "Retrieve details of a specific merchant by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "merchants"
                ],
                "summary": "Get Merchant Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Merchant UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Merchant"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Merchant not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/bymerchant/{merchantID}": {
            "get": {
                "description": "Retrieve all transactions belonging to a specific merchant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List transactions by Merchant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Merchant UUID",
                        "name": "merchantID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/new": {
            "post": {
                "description": "Calculates commission and fees, saves to DB, and returns the created transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "Transaction Request",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler.createTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/transactions": {
            "get": {
                "description": "Retrieve all transactions in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List all transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Transaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "description": "Retrieve a specific transaction details by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get a transaction by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_CardenalDex_crudprotec_internal_entitys.Transaction"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_CardenalDex_crudprotec_internal_entitys.Merchant": {
            "type": "object",
            "properties": {
                "businessID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_CardenalDex_crudprotec_internal_entitys.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Value in cents (200.00 -\u003e 20000)",
                    "type": "integer",
                    "format": "int64"
                },
                "commission": {
                    "description": "in cents",
                    "type": "integer",
                    "format": "int64"
                },
                "deletedAt": {
                    "type": "string"
                },
                "fee": {
                    "description": "in centi% 5.5=550",
                    "type": "integer",
                    "format": "int64"
                },
                "id": {
                    "type": "string"
                },
                "merchantID": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapter_handler.createMerchantRequest": {
            "type": "object",
            "required": [
                "business_id"
            ],
            "properties": {
                "business_id": {
                    "type": "string"
                }
            }
        },
        "internal_adapter_handler.createTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "merchant_id"
            ],
            "properties": {
                "amount": {
                    "description": "Input as float for user friendliness",
                    "type": "number"
                },
                "merchant_id": {
                    "type": "string"
                }
            }
        },
...
    }
}